Fix compilation warnings

+++ libyaul/lib/string/strnlen.c
@@ -3,5 +3,6 @@
 size_t strnlen(const char *s, size_t n)
 {
         const char *p = memchr(s, 0, n);
-        return p ? p-s : n;
+
+        return p ? (size_t)(p-s) : n;
 }
